+++ packages/entities/entities-certificates/src/components/CertificateList.vue
@@ -509,7 +509,7 @@ watch(fetcherState, (state) => {
 const emptyStateOptions = ref<EmptyStateOptions>({
   ctaPath: props.config.createRoute,
   ctaText: undefined,
-  message: t('certificates.list.empty_state.description'),
+  message: `${t('certificates.list.empty_state.description')}${props.config.additionMessageForEmptyState ? ` ${props.config.additionMessageForEmptyState}` : ''}`,
   title: t('certificates.title'),
 })