+++ packages/entities/entities-consumer-groups/sandbox/pages/ConsumerGroupConfigCardPage.vue
@@ -2,7 +2,6 @@
   <h2>Konnect API</h2>
   <ConsumerGroupConfigCard
     :config="konnectConfig"
-    @copy:success="onCopy"
     @fetch:error="onError"
     @fetch:success="onSuccess"
   />
@@ -10,7 +9,6 @@
   <h2>Kong Manager API</h2>
   <ConsumerGroupConfigCard
     :config="kongManagerConfig"
-    @copy:success="onCopy"
     @fetch:error="onError"
     @fetch:success="onSuccess"
   />
@@ -51,8 +49,5 @@ const onError = (error: AxiosError) => {
 }
 const onSuccess = (payload: Record<string, any>) => {
   console.log('fetch:success', payload)
-}
-const onCopy = (payload: Record<string, any>) => {
-  console.log('copy:success', payload)
 }
 </script>