+++ packages/entities/entities-plugins/src/components/PluginList.vue
@@ -751,7 +751,7 @@ watch(fetcherState, (state) => {
 const emptyStateOptions = ref<EmptyStateOptions>({
   ctaPath: props.config.createRoute,
   ctaText: undefined,
-  message: t('plugins.list.empty_state.description'),
+  message: `${t('plugins.list.empty_state.description')}${props.config.additionMessageForEmptyState ? ` ${props.config.additionMessageForEmptyState}` : ''}`,
   title: t('plugins.title'),
 })