+++ packages/entities/entities-snis/src/components/SniList.vue
@@ -307,7 +307,7 @@ const errorMessage = ref<TableErrorMessage>(null)
 const emptyStateOptions = ref<EmptyStateOptions>({
   ctaPath: props.config.createRoute,
   ctaText: undefined,
-  message: t('snis.list.empty_state.description'),
+  message: `${t('snis.list.empty_state.description')}${props.config.additionMessageForEmptyState ? ` ${props.config.additionMessageForEmptyState}` : ''}`,
   title: t('snis.title'),
 })