+++ packages/entities/entities-vaults/sandbox/pages/VaultConfigCardPage.vue
@@ -2,7 +2,6 @@
   <h2>Konnect API</h2>
   <VaultConfigCard
     :config="konnectConfig"
-    @copy:success="onCopy"
     @fetch:error="onError"
     @fetch:success="onSuccess"
   />
@@ -10,7 +9,6 @@
   <h2>Kong Manager</h2>
   <VaultConfigCard
     :config="kongManagerConfig"
-    @copy:success="onCopy"
     @fetch:error="onError"
     @fetch:success="onSuccess"
   />
@@ -50,8 +48,5 @@ const onError = (error: AxiosError) => {
 }
 const onSuccess = (payload: Record<string, any>) => {
   console.log('fetch:success', payload)
-}
-const onCopy = (payload: Record<string, any>) => {
-  console.log('copy:success', payload)
 }
 </script>